Move AlreadyLockedException to ids.plugin.
