+++ pom.xml
@@ -84,7 +84,7 @@
 		<dependency>
 			<groupId>org.icatproject</groupId>
 			<artifactId>ids.plugin</artifactId>
-			<version>1.3.1</version>
+			<version>1.3.2-SNAPSHOT</version>
 		</dependency>
 
 		<dependency>
